% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barwidth.R
\name{barwidth}
\alias{barwidth}
\title{Barwidth}
\usage{
barwidth(x)
}
\arguments{
\item{x}{DNAStringSet}
}
\value{

}
\description{
Calculates witdth and length of names.arg for making an
alphabetByCycle bargraph
}
\examples{
barwidth(test.dss)

}
