% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counter}
\usage{
counter(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
count table
}
\description{
Generate count tables of sRNAs per gene from trans.out files
}
\examples{
a <-counter(test.trans)
head(a)
}
